requirements.txt
frida-tools>=16.0.0,<17.0.0
